+++ lib/market_hub.rb
@@ -9,12 +9,13 @@ require 'market_hub/configuration'
 require 'market_hub/exception'
 require 'market_hub/version'
 
+require 'market_hub/utils/error'
 require 'market_hub/utils/http'
 require 'market_hub/utils/o_auth_20'
+require 'market_hub/utils/validation'
 
 require 'market_hub/models/base'
 require 'market_hub/models/mercado_livre/item'
-require 'market_hub/models/mercado_livre/order'
 
 require 'market_hub/api/base'
 require 'market_hub/api/mercado_livre/client'