+++ test/market_hub/models/mercado_livre/test_item.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class MarketHub::Models::MercadoLivre::TestItem < Minitest::Test
+end